Consolidate plugin settings interfaces into a single TypeScript module
