+++ npm-packages/meteor-vite/src/plugin/MeteorStubs.ts
@@ -8,9 +8,8 @@ import { createErrorHandler } from '../error/ErrorHandler';
 import { MeteorViteError } from '../error/MeteorViteError';
 import MeteorPackage from '../meteor/package/components/MeteorPackage';
 import { stubTemplate } from '../meteor/package/StubTemplate';
-import { MeteorViteConfig } from '../MeteorViteConfig';
+import { MeteorViteConfig, type PluginSettings } from '../MeteorViteConfig';
 import ViteLoadRequest from '../ViteLoadRequest';
-import { PluginSettings } from './Meteor';
 
 export const MeteorStubs = setupPlugin(async () => {
     return {
@@ -145,104 +144,5 @@ function setupPlugin<Context extends ViteLoadRequest>(setup: () => Promise<{
 }
 
 
-export interface MeteorStubsSettings {
-    
-    meteor: {
-        /**
-         * Path to Meteor's internal package cache.
-         * This can change independently of the isopack path depending on whether we're building for production or
-         * serving up the dev server.
-         *
-         * @example {@link /examples/vue/.meteor/local/build/programs/web.browser/packages}
-         */
-        packagePath: string;
-        
-        /**
-         * Path to Meteor's local Isopacks store. Used to determine where a package's mainModule is located and whether
-         * the package has lazy-loaded modules. During production builds this would be pulled from a temporary
-         * Meteor build, so that we have solid metadata to use when creating Meteor package stubs.
-         *
-         * @example {@link /examples/vue/.meteor/local/isopacks/}
-         */
-        isopackPath: string;
-        
-        /**
-         * Path to the current user's Meteor package cache. (e.g. /home/john/.meteor/packages)
-         * This is used to build up a fallback path for isopack manifests.
-         *
-         * Some packages, like `react-meteor-data` do not emit a isopack metadata file within the current project's
-         * .meteor/local directory. So we have to resort to pulling in Isopack metadata from the `meteor-tool` cache.
-         *
-         * @example `react-meteor-data` path
-         * /home/john/.meteor/packages/react-meteor-data/2.7.2/web.browser.json
-         */
-        globalMeteorPackagesDir?: string;
-    }
-    
-    /**
-     * Full content of the user's Meteor project package.json.
-     * Like the one found in {@link /examples/vue/package.json}
-     */
-    packageJson?: ProjectJson;
-    
-    /**
-     * Alternatively, a path to a package.json file can be supplied.
-     */
-    packageJsonPath?: string;
-    
-    /**
-     * Enabling debug mode will write all input and output files to a `.meteor-vite` directory in the Meteor project's
-     * root. Handy for quickly assessing how things are being formatted, or for use in writing up new test cases for
-     * meteor-vite.
-     */
-    debug?: boolean;
-    
-}
-
 type ResolvedPluginConfig = Required<PluginSettings>;
 
-/**
- * The user's Meteor project package.json content.
- * todo: expand types
- */
-export type ProjectJson = {
-    name: string;
-    dependencies: {
-        'meteor-vite'?: string;
-    }
-    devDependencies: {
-        'meteor-vite'?: string;
-    }
-    meteor: {
-        mainModule: {
-            client: string;
-        },
-        /**
-         * @deprecated Use meteor.vite.configFile instead.
-         * See {@link https://github.com/JorgenVatle/meteor-vite?tab=readme-ov-file#configuration configuration} for
-         * example.
-         */
-        viteConfig?: string;
-        vite?: {
-            /**
-             * Specifies an alternative path to the project's Vite config
-             */
-            configFile?: string;
-            
-            /**
-             * Remove or replace Meteor packages when preparing the intermediary production build.
-             * Does not affect your final production bundle. It's only used as a temporary build step.
-             */
-            replacePackages?: {
-                startsWith: string; // Match any Meteor package name that starts with the provided string.
-                replaceWith: string; // Replace matching packages with the provided string.
-            }[];
-            
-            /**
-             * Override the directory path used for preparing the Vite production bundle.
-             * Might be useful if the automatically generated file path is inaccessible in your operating system
-             */
-            tempBuildDir?: string;
-        }
-    }
-}
